refactor: remove sqlite oauth token store
