+++ packages/browseros-agent/apps/server/src/api/server.ts
@@ -18,7 +18,6 @@ import type { ContentfulStatusCode } from 'hono/utils/http-status'
 import { HttpAgentError } from '../agent/errors'
 import { INLINED_ENV } from '../env'
 import { KlavisClient } from '../lib/clients/klavis/klavis-client'
-import type { OAuthTokenManager } from '../lib/clients/oauth/token-manager'
 import { logger } from '../lib/logger'
 import { Sentry } from '../lib/sentry'
 import { getLimaHomeDir, resolveBundledLimactl, VM_NAME } from '../lib/vm'
@@ -31,7 +30,6 @@ import { createKlavisRoutes } from './routes/klavis'
 import { createMcpRoutes } from './routes/mcp'
 import { createMemoryRoutes } from './routes/memory'
 import { createMonitoringRoutes } from './routes/monitoring'
-import { createOAuthRoutes } from './routes/oauth'
 import { createOpenClawRoutes } from './routes/openclaw'
 import { createProviderRoutes } from './routes/provider'
 import { createRefinePromptRoutes } from './routes/refine-prompt'
@@ -87,8 +85,6 @@ export async function createHttpServer(config: HttpServerConfig) {
   } = config
 
   const { onShutdown } = config
-
-  const tokenManager: OAuthTokenManager | null = null
 
   const aclPolicyService = new GlobalAclPolicyService()
   await aclPolicyService.load()
@@ -167,7 +163,6 @@ export async function createHttpServer(config: HttpServerConfig) {
       '/shutdown',
       createShutdownRoute({
         onShutdown: () => {
-          tokenManager?.stopCallbackServer()
           stopKlavisBackground()
           klavisRef.handle?.close().catch((err) =>
             logger.warn('Failed to close Klavis proxy transport', {
@@ -188,11 +183,9 @@ export async function createHttpServer(config: HttpServerConfig) {
     .route('/refine-prompt', createRefinePromptRoutes({ browserosId }))
     .route(
       '/oauth',
-      tokenManager
-        ? createOAuthRoutes({ tokenManager })
-        : new Hono().all('/*', (c) =>
-            c.json({ error: 'OAuth not available' }, 503),
-          ),
+      new Hono().all('/*', (c) =>
+        c.json({ error: 'OAuth not available' }, 503),
+      ),
     )
     .route('/klavis', createKlavisRoutes({ browserosId: browserosId || '' }))
     .route(
